refactor(client): 拆分 App.jsx 为组件 + hooks + utils
将原 2093 行的 App.jsx 拆成 26 个独立模块（组件 / hooks / utils），
App.jsx 从 2093 行降到约 891 行，只保留顶层装配与跨模块协调逻辑。
重构包含此前遗漏的接线收尾：补齐 useMemo / PANEL_SIZE_KEY / CopyButton
import，调用 useSessionHandlers 生成 handleCreateSession 等 5 个 handler
和 setChatModel；删除未使用的 submitAgentApproval 引用。
