+++ client/src/components/markdown/CodeBlock.jsx
@@ -1,0 +1,38 @@
+import { useEffect, useRef } from 'react';
+import hljs from 'highlight.js/lib/core';
+import javascript from 'highlight.js/lib/languages/javascript';
+import python from 'highlight.js/lib/languages/python';
+import css from 'highlight.js/lib/languages/css';
+import xml from 'highlight.js/lib/languages/xml';
+import json from 'highlight.js/lib/languages/json';
+import bash from 'highlight.js/lib/languages/bash';
+import sql from 'highlight.js/lib/languages/sql';
+import 'highlight.js/styles/github.css';
+
+hljs.registerLanguage('javascript', javascript);
+hljs.registerLanguage('js', javascript);
+hljs.registerLanguage('python', python);
+hljs.registerLanguage('py', python);
+hljs.registerLanguage('css', css);
+hljs.registerLanguage('html', xml);
+hljs.registerLanguage('xml', xml);
+hljs.registerLanguage('json', json);
+hljs.registerLanguage('bash', bash);
+hljs.registerLanguage('sh', bash);
+hljs.registerLanguage('sql', sql);
+
+export function CodeBlock({ language, children, ...props }) {
+  const ref = useRef(null);
+  useEffect(() => {
+    if (ref.current && language && hljs.getLanguage(language)) {
+      hljs.highlightElement(ref.current);
+    }
+  }, [children, language]);
+  return (
+    <pre className="code-block" {...props}>
+      <code ref={ref} className={language ? `hljs language-${language}` : ''}>
+        {children}
+      </code>
+    </pre>
+  );
+}
