+++ client/src/utils/constants.js
@@ -1,0 +1,9 @@
+export const PHONE_BREAKPOINT = 640;
+export const TABLET_BREAKPOINT = 768;
+export const DOCKED_LAYOUT_BREAKPOINT = 1100;
+export const APP_BG_COLOR = '#f5f5fa';
+export const APP_SURFACE_COLOR = '#ffffff';
+
+export const PANEL_MIN = 280;
+export const PANEL_MAX_RATIO = 0.7;
+export const PANEL_SIZE_KEY = 'agent_panel_width';
