+++ client/src/utils/format.js
@@ -1,0 +1,10 @@
+export function formatMsgTime(ts) {
+  if (!ts) return '';
+  return new Intl.DateTimeFormat('zh-CN', {
+    hour: '2-digit',
+    minute: '2-digit',
+    second: '2-digit',
+    hour12: false,
+    timeZone: 'Asia/Shanghai',
+  }).format(ts);
+}
